fix: migrate auth code from user to auth structure
