+++ internal/repositories/auth.go
@@ -1,0 +1,8 @@
+package repositories
+
+import "LeaseEase/internal/models"
+
+type AuthRepository interface {
+	CreateUser(user *models.User) error
+	GetUserByEmail(email string) (*models.User, error)
+}
