+++ internal/repositories/auth_repository.go
@@ -1,0 +1,27 @@
+package repositories
+
+import (
+	"LeaseEase/internal/models"
+
+	"gorm.io/gorm"
+)
+
+type authRepository struct {
+	db *gorm.DB
+}
+
+func NewAuthRepository(db *gorm.DB) AuthRepository {
+	return &authRepository{
+		db: db,
+	}
+}
+
+func (r *authRepository) CreateUser(user *models.User) error {
+	return r.db.Create(user).Error
+}
+
+func (r *authRepository) GetUserByEmail(email string) (*models.User, error) {
+	var user models.User
+	err := r.db.Where("email = ?", email).First(&user).Error
+	return &user, err
+}
