+++ internal/repositories/repository.go
@@ -3,4 +3,5 @@ package repositories
 type Repository interface {
 	User() UserRepository
 	Property() PropertyRepository
+	Auth() AuthRepository
 }
