+++ internal/repositories/repository_impl.go
@@ -9,12 +9,14 @@ import (
 type repository struct {
 	UserRepository     UserRepository
 	PropertyRepository PropertyRepository
+	AuthRepository AuthRepository
 }
 
 func NewRepository(cfg *config.DBConfig, db *gorm.DB) Repository {
 	return &repository{
 		UserRepository:     NewUserRepository(db),
 		PropertyRepository: NewPropertyRepository(db),
+		AuthRepository:     NewAuthRepository(db),
 	}
 }
 
@@ -24,4 +26,8 @@ func (r *repository) User() UserRepository {
 
 func (r *repository) Property() PropertyRepository {
 	return r.PropertyRepository
+}
+
+func (r *repository) Auth() AuthRepository {
+	return r.AuthRepository
 }
