+++ internal/services/auth.go
@@ -1,0 +1,8 @@
+package services
+
+import "LeaseEase/internal/dtos"
+
+type AuthService interface {
+	Register(registerDTO *dtos.RegisterDTO) error
+	Login(loginDTO *dtos.LoginDTO) (string, error)
+}
