+++ internal/services/service.go
@@ -3,4 +3,5 @@ package services
 type Service interface {
 	User() UserService
 	Property() PropertyService
+	Auth() AuthService
 }
