+++ internal/services/service_impl.go
@@ -5,12 +5,14 @@ import "LeaseEase/internal/repositories"
 type service struct {
 	UserService UserService
 	PropertyService PropertyService
+	AuthService AuthService
 }
 
 func NewService(repo repositories.Repository) Service {
 	return &service{
 		UserService: NewUserService(repo.User()),
 		PropertyService: NewPropertyService(repo.Property()),
+		AuthService: NewAuthService(repo.Auth()),
 	}
 }
 
@@ -22,3 +24,6 @@ func (s *service) Property() PropertyService {
 	return s.PropertyService
 }
 
+func (s *service) Auth() AuthService {
+	return s.AuthService
+}
